+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 * Add missing `paymentMethodToken` search to `CreditCardVerificationSearchRequest`
+* Add `recurringCustomerConsent` and `recurringMaxAmount` parameters to `authenticationInsightOptions` in `PaymentMethodNonce.create()`
+
 
 ## 2.108.0
 * Fix null pointer exception when 3DS lookup returns a `422`